Clean up shared code a bit
