+++ lib/global/event-handlers/onGroupPointerLeave.ts
@@ -1,17 +1,27 @@
 import { read } from "../mutableState";
-import { onWindowPointerMove } from "./onWindowPointerMove";
+import { updateActiveHitRegions } from "../utils/updateActiveHitRegion";
 
 export function onGroupPointerLeave(event: PointerEvent) {
-  const { interactionState } = read();
+  if (event.defaultPrevented) {
+    return;
+  } else if (event.relatedTarget !== null) {
+    return;
+  }
 
+  const { interactionState, mountedGroups } = read();
+
+  // The "pointerleave" event is not reliably triggered when the pointer exits a window or iframe
+  // To account for this, we listen for "pointerleave" events on the Group element itself
   switch (interactionState.state) {
     case "active": {
       interactionState.hitRegions.forEach((hitRegion) => {
         if (event.currentTarget === hitRegion.group.element) {
-          // The "pointerleave" event is not reliably triggered when the pointer exits a window or iframe
-          // To account for this, we listen for "pointerleave" events on the Group element itself
-          // The easiest way to handle these is to reuse the same logic as the Window "pointermove" event
-          onWindowPointerMove(event);
+          updateActiveHitRegions({
+            event,
+            hitRegions: interactionState.hitRegions,
+            initialLayoutMap: interactionState.initialLayoutMap,
+            mountedGroups
+          });
         }
       });
     }
